# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Election Integrity Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Analysis status of the acta
 */
export type ActaStatus = (typeof ActaStatus)[keyof typeof ActaStatus];

export const ActaStatus = {
  valid: "valid",
  suspicious: "suspicious",
  inconsistent: "inconsistent",
} as const;

export interface Acta {
  id: number;
  mesa_id: string;
  /**
   * Colombian department name
   * @nullable
   */
  departamento?: string | null;
  /**
   * Colombian municipality name
   * @nullable
   */
  municipio?: string | null;
  image_path: string;
  /** SHA256 hash of the image */
  image_hash: string;
  /**
   * OCR extracted text from the image
   * @nullable
   */
  ocr_result?: string | null;
  /** Analysis status of the acta */
  status: ActaStatus;
  /**
   * Hash of the previous block in chain
   * @nullable
   */
  prev_hash?: string | null;
  /** SHA256 of prev_hash + image_hash + timestamp */
  block_hash: string;
  timestamp: string;
}

export interface DashboardStats {
  total: number;
  valid: number;
  suspicious: number;
  inconsistent: number;
}

export interface BlockchainVerification {
  is_valid: boolean;
  block_count: number;
  message: string;
  /** @nullable */
  broken_at_block?: number | null;
}

export interface Block {
  id: number;
  mesa_id: string;
  block_hash: string;
  /** @nullable */
  prev_hash?: string | null;
  image_hash: string;
  timestamp: string;
  status: string;
}

export type UploadActaBody = {
  /** Electoral table identifier */
  mesa_id: string;
};