# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Election Integrity Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /actas/{id}:
    get:
      operationId: getActa
      tags: [actas]
      summary: Get acta detail
      description: Returns full detail of a single acta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Acta detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acta"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /actas/upload:
    post:
      operationId: uploadActa
      tags: [actas]
      summary: Upload an acta image
      description: Upload an electoral record image (multipart/form-data), runs OCR, generates SHA256 hash, and adds to blockchain. Fields mesa_id (string) and image (file) required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mesa_id
              properties:
                mesa_id:
                  type: string
                  description: Electoral table identifier
      responses:
        "201":
          description: Acta uploaded and analyzed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acta"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Acta:
      type: object
      properties:
        id:
          type: integer
        mesa_id:
          type: string
        departamento:
          type: ["string", "null"]
          description: Colombian department name
        municipio:
          type: ["string", "null"]
          description: Colombian municipality name
        image_path:
          type: string
        image_hash:
          type: string
          description: SHA256 hash of the image
        ocr_result:
          type: ["string", "null"]
          description: OCR extracted text from the image
        status:
          type: string
          enum: [valid, suspicious, inconsistent]
          description: Analysis status of the acta
        prev_hash:
          type: ["string", "null"]
          description: Hash of the previous block in chain
        block_hash:
          type: string
          description: SHA256 of prev_hash + image_hash + timestamp
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - mesa_id
        - image_path
        - image_hash
        - status
        - block_hash
        - timestamp
...