# lib/api-zod/src/generated/types/acta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Election Integrity Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActaStatus } from "./actaStatus";

export interface Acta {
  id: number;
  mesa_id: string;
  /**
   * Colombian department name
   * @nullable
   */
  departamento?: string | null;
  /**
   * Colombian municipality name
   * @nullable
   */
  municipio?: string | null;
  image_path: string;
  /** SHA256 hash of the image */
  image_hash: string;
  /**
   * OCR extracted text from the image
   * @nullable
   */
  ocr_result?: string | null;
  /** Analysis status of the acta */
  status: ActaStatus;
  /**
   * Hash of the previous block in chain
   * @nullable
   */
  prev_hash?: string | null;
  /** SHA256 of prev_hash + image_hash + timestamp */
  block_hash: string;
  timestamp: Date;
}